cross_platform_app/backend/requirements.txt
flask
newspaper3k
gunicorn